feat: refactor sidebar component and add menu list constants
